pyproject.toml
[project]
name = "nba-predictor"
version = "0.1.0"
description = "Automated ML pipeline predicting NBA players' next-game points, built on Supabase + XGBoost + Streamlit."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nba-api>=1.11.4",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "tenacity>=9.1.4",
    "xgboost>=3.2.0",
]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["notebooks", ".venv", "artifacts"]

[tool.ruff.lint]
# Sensible default rule set: pyflakes (F), pycodestyle (E/W), isort (I),
# pyupgrade (UP), bugbear (B), comprehensions (C4), simplify (SIM).
# Keep it tight — lint noise erodes trust in the signal.
select = ["E", "W", "F", "I", "UP", "B", "C4", "SIM"]
ignore = [
    "E501",  # line too long — formatter owns wrapping
    "B008",  # function call in default arg — common/safe for FastAPI/Typer-style code
]

[tool.ruff.lint.per-file-ignores]
# Tests import fixtures with patterns ruff doesn't like; keep them readable.
"tests/**" = ["B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra --strict-markers"
